+++ packages/json-rpc/json-rpc-provider-proxy/src/json-rpc-message.ts
@@ -1,0 +1,5 @@
+export const jsonRpcMsg = <T extends {}>(msg: T) =>
+  JSON.stringify({
+    jsonrpc: "2.0",
+    ...msg,
+  })