+++ packages/json-rpc/json-rpc-provider-proxy/src/subscription-manager/subscription-manager.ts
@@ -1,10 +1,10 @@
-import type {
-  RequestId,
-  SubscriptionId,
-  SubscriptionLogic,
+import {
+  MessageType,
+  type RequestId,
+  type SubscriptionId,
+  type SubscriptionLogic,
 } from "../internal-types"
 import { chainHeadFollow } from "./chainHeadFollow"
-import { txSubmitAndWatch } from "./transaction-submit-watch"
 
 export const addSubscription = ({
   onSent,
@@ -25,23 +25,22 @@ export const addSubscription = ({
   return {
     onSent(parsed: any) {
       const result = onSent(parsed)
-      if (!result) return
-      if (result.type === "subscribe") {
-        preActive.set(result.id, result.onRes)
-      } else {
-        active.delete(result.id)
+      if (result) {
+        if (result.type === MessageType.subscribe)
+          preActive.set(result.id, result.onRes)
+        else active.delete(result.id)
       }
     },
     onResponse(parsed: any) {
       const match = preActive.get(parsed.id)?.(parsed)
-      if (!match) return
-      preActive.delete(parsed.id)
-      active.add(match.id)
+      if (match) {
+        preActive.delete(parsed.id)
+        active.add(match.id)
+      }
     },
     onNotifiaction(parsed: any) {
       const result = onNotification(parsed)
-      if (!result) return
-      active.delete(result.id)
+      if (result) active.delete(result.id)
     },
     onDisconnect,
     onAbort() {
@@ -52,32 +51,5 @@ export const addSubscription = ({
   }
 }
 
-export const getSubscriptionManager = (onMessage: (msg: string) => void) => {
-  const subscriptions = [chainHeadFollow, txSubmitAndWatch].map((logic) =>
-    addSubscription(logic(onMessage)),
-  )
-
-  return {
-    onSent(parsed: any) {
-      subscriptions.forEach((s) => {
-        s.onSent(parsed)
-      })
-    },
-    onResponse(parsed: any) {
-      subscriptions.forEach((s) => {
-        s.onResponse(parsed)
-      })
-    },
-    onNotifiaction(parsed: any) {
-      subscriptions.forEach((s) => {
-        s.onNotifiaction(parsed)
-      })
-    },
-    onDisconnect() {
-      subscriptions.forEach((s) => s.onDisconnect())
-    },
-    onAbort() {
-      subscriptions.forEach((s) => s.onAbort())
-    },
-  }
-}
+export const getSubscriptionManager = (onMessage: (msg: string) => void) =>
+  addSubscription(chainHeadFollow(onMessage))