+++ packages/known-chains/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Add Paseo testnet
+
+### Fixed
+
+- Update `lightSyncState` and `bootnodes`
+
 ## 0.1.6 - 2024-06-17
 
 ### Fixed